fix protocol config and expose cipher options for sshe layers
